+++ NEWS.md
@@ -9,6 +9,7 @@
 
 - Internal function `is_doc_open()` replaces `is_office_doc_edited()` to check if a document is open on Windows. This solves the issue where RStudio crashed while trying to write to an open Office document.
 - Add option `preserve` to preserve cell line breaks to `docx_summary()`.
+- fix `ph_hyperlink()` when several hyperlinks are inside one slide and refactored `ph_remove()`.
 
 # officer 0.6.2